[illicit] move AnonRc to its own module for readability.
